+++ packages/bundled-features/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @cosmicdrift/kumiko-bundled-features
 
+## 0.159.1
+
+### Patch Changes
+
+- Updated dependencies [6d37eb5]
+  - @cosmicdrift/kumiko-framework@0.159.1
+  - @cosmicdrift/kumiko-headless@0.159.1
+  - @cosmicdrift/kumiko-renderer@0.159.1
+  - @cosmicdrift/kumiko-dispatcher-live@0.159.1
+  - @cosmicdrift/kumiko-renderer-web@0.159.1
+
 ## 1.0.0
 
 ### Major Changes